Allocate 901,000 - 929,999 for new desktop user UID/GIDs
To keep desktop users from overlapping with Android's new
multi-user device profiles, we use the offset 900,000 for
any new user IDs. This is based on the fictionary "u09" since
Android's UserManagerService allocates secondary owners starting
at u10.

The offset for any user uM is calculated as 100,000 * M. This is
based on the hardcoded AID_USER offset of 100,000 found in
/system/core/include/private/android_filesystem_config.h