Fully automate Maru Desktop build
After a single dpkg -i and apt-get install -f everything
for Maru Desktop is configured.

There is one weird thing though--you have to login twice for
the desktop icons to have the proper spacing underneath the
top panel. On the first boot they are too high, as if the panel
did not add the extra spacing. On the second boot they are properly
padded.

[TODO]
Package removal and upgrade needs to be tested.