Change Camera desktop shortcut to overall Phone shortcut
Unfortunately the /sdcard/DCIM/Camera folder is only created after
the user launches the Camera app for the first time so we can't
rely on it always being there.

Simple solution is to link to the top of the mount point. Whenever
the user gets to open the Camera app for the first time, the desktop
will see this Camera/ show up under DCIM/.

[Testing]
  - link functions without error on first boot before user opens
    up Camera app
  - /mnt/Phone/DCIM/Camera shows up in real-time as soon as the user
    starts the Camera app