add blueprint-specific options
Blueprint-specific options are necessary so that blueprint scripts can
customize the build process. This is useful for passing things like the
release to use as a base for the image, architecture, etc. Some of these
options may or may not be relevant to each blueprint, so let the
blueprint itself define it's own list of options.

In order to do this, we switch to using getopt to handle both short and
long options in a concise way.

Also a few other improvements:
    - ensure that the root account password is set to "root"
    - clean up debpkg artifacts
    - use the dynamic APT mirror httpredir.debian.org/debian