debian: Add --minimal option to minimize image size
There is a new -m,--minimal option for the debian blueprint that shrinks
the image size from 450MB down to ~330MB. This is useful for devices
with limited system partition space.

The real motivation for this is the Nexus 7, which only has a 650MB
system partition so space is tight and it needs a desktop image that is
< 350MB.

Right now, the minimal build just drops the Recommended packages from
the installation. The main maru package had to add some stuff to Depends
to get X / LightDM / XFCE to launch properly, most importantly xinit and
xserver-xorg. tango-icon-theme is also explicitly included so the
minimal image build has a consistent look with the full image, because
it is only listed as a recommends for xfce4. Basically, the maru Depends
section should list the packages that MUST be installed no matter what
for Maru Desktop, so that their recommended dependencies can be dropped
with no major side-effects.

Also--included a trap cleanup function on EXIT so this script can clean
up after itself no matter what! This includes Ctrl-C, very convenient.

Signed-off-by: Preetam D'Souza <preetamjdsouza@gmail.com>