debian: Update LXC template file
This is based on the latest lxc-debian.in template on LXC's stable-1.0
branch with a few small mods:

 - Use qemu-debootstrap for cross-bootstrapping
 - Remove the systemd masking of udev
 - Fill in the automake template variables with the defaults

This update is important for the upcoming move to Debian 9 / stretch as
the default release for Maru. There are a few changes to some of the
core pkg install logic like explicitly including an init system that is
necessary on stretch. This also includes a few improvements like using
the Debian security archives by default which is good.

I have successfully tested this with both jessie and stretch container
creation so we can continue to generate jessie containers if needed.

Signed-off-by: Preetam D'Souza <preetamjdsouza@gmail.com>