debian: Ignore device touchscreen
On stretch, libinput is now the default input handling system for X and
it configures the actual device touchscreen itself as an input device
out-of-the-box. This causes issues because any time the user interacts
with the device screen, it also forwards those events to the desktop.
This change updates the X config file to ignore any touchscreens.

As a side note: this could be really awesome for using the device itself
as a pointer device, which has been requested many times by Maru users.
We will need a way to dynamically toggle touchscreen input to make this
happen.

Signed-off-by: Preetam D'Souza <preetamjdsouza@gmail.com>