Upgrade to buster
Support building Debian 10 Buster containers.

Building armhf Buster containers on QEMU is a bit of a pain since there
are some overflow issues with building 32-bit on a 64-bit that shows up
with the the version of glibc shipped in Buster. Fortunately, we can
work around this by using the 32-bit version of qemu-user-static.

One notable change is that we now download the base rootfs directly from
LXC's servers instead of running debootstrap from scratch - this is made
possible in LXC 3.0 with the 'download' template. This speeds up the
first build a lot.

Summary of changes:

* Update container bootstrapping to use new LXC 3.0 download template
* Use qemu-user-static:i386 to work around kernel/QEMU issues building
32-bit armhf images
* Mount essential filesystems like /proc and /dev before chroot'ing
* Remove unnecessary LXC template script

Tested on a Nexus 5.

Signed-off-by: Preetam D'Souza <preetamjdsouza@gmail.com>